manas_wellness/requirements.txt
# 🧠 Manas: Youth Mental Wellness Platform
# Complete Production Dependencies for GenAI Hackathon

# ===== WEB FRAMEWORK =====
Flask>=3.0.0
Werkzeug>=3.0.1
gunicorn>=21.2.0  # Production WSGI server

# ===== GOOGLE AI & CLOUD SERVICES =====
google-generativeai>=0.8.0  # Gemini API (already working)
google-cloud-aiplatform>=1.40.0
google-cloud-speech>=2.21.0  # Voice-to-text
google-cloud-texttospeech>=2.16.3  # Text-to-speech
google-cloud-translate>=3.12.1  # Multi-language support
google-cloud-vision>=3.4.4  # Facial emotion detection
google-cloud-storage>=2.10.0  # File storage

# ===== GOOGLE FIT INTEGRATION =====
google-auth>=2.23.0  # Google authentication
google-auth-oauthlib>=1.1.0  # OAuth 2.0 flow
google-auth-httplib2>=0.2.0  # HTTP library for Google APIs
google-api-python-client>=2.100.0  # Google API client

# ===== DATABASE SERVICES =====
# Current SQLite (included in Python)
# Production databases:
pymongo>=4.6.0  # MongoDB integration
supabase>=2.3.0  # Real-time database
redis>=5.0.1  # Caching and sessions

# ===== MUSIC THERAPY INTEGRATION =====
spotipy>=2.22.1  # Spotify API client
mutagen>=1.47.0  # Music metadata

# ===== REAL-TIME FEATURES =====
flask-socketio>=5.3.0  # WebSocket support
python-socketio>=5.10.0  # Socket.IO client
eventlet>=0.33.3  # Async networking

# ===== SMS/COMMUNICATION =====
twilio>=8.10.0  # Crisis SMS alerts

# ===== COMPUTER VISION & AI =====
opencv-python>=4.9.0
mediapipe>=0.10.21  # Eye tracking, gesture recognition
Pillow>=10.2.0

# ===== AUDIO PROCESSING =====
librosa>=0.10.1  # Audio analysis
soundfile>=0.12.1
# pyaudio>=0.2.11  # Commented - can cause install issues


# ===== DATA PROCESSING & ML =====
numpy>=1.24.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.22.0

# ===== TEXT PROCESSING & NLP =====
nltk>=3.8.0
textblob>=0.17.0

# ===== HTTP & API CLIENTS =====
requests>=2.31.0
urllib3>=2.1.0
httpx>=0.25.0  # Async HTTP client

# ===== SECURITY & AUTHENTICATION =====
flask-cors>=4.0.0
flask-limiter>=3.5.0
cryptography>=41.0.0
bcrypt>=4.1.0
PyJWT>=2.8.0  # JWT tokens
flask-session>=0.5.0

# ===== CONFIGURATION & ENVIRONMENT =====
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ===== UTILITIES =====
python-dateutil>=2.8.2
jsonschema>=4.20.0
tqdm>=4.66.0  # Progress bars
psutil>=5.9.0  # System monitoring

# ===== LOGGING & MONITORING =====
colorlog>=6.8.0
sentry-sdk>=1.38.0  # Error tracking (optional)

# ===== BACKGROUND TASKS (Optional) =====
# celery>=5.3.0  # Uncomment if using background tasks

# ===== DEVELOPMENT TOOLS (Install separately if needed) =====
# pytest>=7.4.0
# pytest-flask>=1.3.0
# black>=23.11.0
# flake8>=6.1.0

# ===== DEPLOYMENT REQUIREMENTS =====
# For production deployment:
# - Docker support: docker>=6.1.0
# - Load balancing: nginx (system package)
# - Process management: supervisor (system package)
# - SSL certificates: certbot (system package)